refactor: modularize gantt hooks and utilities
