+++ src/components/Gantt/hooks/useGantt.ts
@@ -1,8 +1,10 @@
-import { useCallback, useMemo, useRef } from "react";
+import { useMemo, useRef } from "react";
 import { getTimeSequence } from "../utils/getTimeSequence";
 import { useElementWidth } from "./useElementWidth";
 import { roundTimeRange } from "../utils/roundTimeRange";
-import { GanttApi, TimeRange, TimeUnit } from "../types";
+import { TimeRange, TimeUnit } from "../types";
+import { useTimeScale } from "./useTimeScale";
+import { useGanttApi } from "./useGanttApi";
 
 type UseGanttProps = {
   timeRange: TimeRange;
@@ -17,7 +19,6 @@ export const useGantt = ({
 }: UseGanttProps) => {
   const ganttRef = useRef<HTMLDivElement | undefined>();
   const canvasRef = useRef<HTMLDivElement | undefined>();
-  const apiRef = useRef<GanttApi | undefined>();
 
   timeRange = useMemo(
     () => roundTimeRange(timeRange, timeUnit),
@@ -26,43 +27,17 @@ export const useGantt = ({
   const { elementWidth: ganttWidth } = useElementWidth({ ref: ganttRef });
   const { elementWidth: canvasWidth } = useElementWidth({ ref: canvasRef });
 
-  const scale = useMemo(
-    () => (timeRange.end - timeRange.start) / canvasWidth,
-    [canvasWidth, timeRange]
-  );
+  const { toPx, toTime } = useTimeScale(timeRange, canvasWidth);
   const timeSequence = useMemo(
     () => getTimeSequence(timeRange, timeUnit),
-    [getTimeSequence, timeRange, timeUnit]
-  );
-
-  const toPx = useCallback(
-    (time: number) => {
-      return (time - timeRange.start) / scale;
-    },
-    [timeRange, scale]
-  );
-
-  const toTime = useCallback(
-    (px: number) => {
-      return px * scale + timeRange.start;
-    },
-    [timeRange, scale]
+    [timeRange, timeUnit]
   );
 
   const visibleCanvasWidth = useMemo(() => {
     return ganttWidth - leftHeaderWidth;
   }, [ganttWidth, leftHeaderWidth]);
 
-  apiRef.current = useMemo(() => {
-    return {
-      scrollToNow: () => {
-        (ganttRef.current as any).scrollLeft = Math.max(
-          0,
-          toPx(Date.now()) - visibleCanvasWidth / 2
-        );
-      },
-    };
-  }, [ganttRef, ganttWidth, leftHeaderWidth, toPx]);
+  const apiRef = useGanttApi(ganttRef as any, visibleCanvasWidth, toPx);
 
   return {
     ganttRef,
