+++ src/components/Gantt/index.ts
@@ -10,8 +10,11 @@ export { TopHeader } from "./components/TopHeader";
 export { TimeDisplay } from "./components/TimeDisplay";
 export { useGantt } from "./hooks/useGantt";
 export { useDraggableScroll } from "./hooks/useDraggableScroll";
+export { useGanttApi } from "./hooks/useGanttApi";
+export { useTimeScale } from "./hooks/useTimeScale";
 export { getSubrows } from "./utils/getSubrows";
 export { getTimeSequence } from "./utils/getTimeSequence";
 export { isCurrent } from "./utils/isCurrent";
 export { roundTimeRange } from "./utils/roundTimeRange";
+export { formatTime } from "./utils/formatTime";
 export { DAY, TimeUnit } from "./types";
